refactor(settings): decompose ToolsSettings into tools/ components
ToolsSettings.tsx was 796 lines — a host + 7 inline sub-components + secret
heuristics. Extract into src/components/settings/tools/ (no behaviour or
styling change; className strings preserved verbatim):

- helpers.ts — HeaderRow type + looksSecret/looksLikeCredential/headersToRows
  + the shared inputCls/labelCls constants (pure, unit-testable).
- HeaderEditor / TestButton / McpAuthButton / McpForm / ServiceForm / ToolRow /
  McpToolList — one file each.

Host ToolsSettings.tsx drops to ~250 lines (store wiring, resolveHeaders +
save handlers, and composition). Only ToolsSettings is imported externally
(settings-view), so extraction is call-site-safe.

Verified: type-check clean; eslint clean; full suite (1128) green.
